feat: move deprecated import RxSwift to its place
