--- p2p_wallet/RxSwift_Deprecated/ObservableType+Extensions.swift
@@ -1,20 +1,0 @@
-//
-//  ObservableType+Extensions.swift
-//  p2p_wallet
-//
-//  Created by Chung Tran on 08/06/2021.
-//
-
-import Foundation
-import RxSwift
-
-extension ObservableType {
-    func withPrevious() -> Observable<(Element?, Element)> {
-        scan([], accumulator: { previous, current in
-            Array(previous + [current]).suffix(2)
-        })
-            .map { arr -> (previous: Element?, current: Element) in
-                (arr.count > 1 ? arr.first : nil, arr.last!)
-            }
-    }
-}
