+++ p2p_wallet/RxSwift_Deprecated/PendingTransaction.swift
@@ -6,7 +6,6 @@
 //
 
 import Foundation
-import RxSwift
 import SolanaSwift
 import TransactionParser
