+++ p2p_wallet/Scenes/Main/History/History.SceneModel.swift
@@ -254,3 +254,14 @@ extension History {
         }
     }
 }
+
+private extension ObservableType {
+    func withPrevious() -> Observable<(Element?, Element)> {
+        scan([], accumulator: { previous, current in
+            Array(previous + [current]).suffix(2)
+        })
+            .map { arr -> (previous: Element?, current: Element) in
+                (arr.count > 1 ? arr.first : nil, arr.last!)
+            }
+    }
+}
