+++ p2p_wallet/Scenes/Main/ReceiveToken/Views/ShowHideButton.swift
